requirements.txt
numpy
polars
scikit-learn
torch
opencv-python
PyYAML
google-genai
pydantic